fix(liveness): remove self-check JMX request on probe
This reverts commit 3c19582259a992e3b41bb3c5a5ffec1e7a83f1a5 (#1606).
