+++ schema-generator/dependency-reduced-pom.xml
@@ -103,7 +103,7 @@
   </build>
   <properties>
     <maven.compiler.release>21</maven.compiler.release>
-    <com.diffplug.spotless.maven.plugin.version>3.5.1</com.diffplug.spotless.maven.plugin.version>
+    <com.diffplug.spotless.maven.plugin.version>3.4.0</com.diffplug.spotless.maven.plugin.version>
     <maven.compiler.source>21</maven.compiler.source>
     <com.mycila.license.maven.plugin.version>5.0.0</com.mycila.license.maven.plugin.version>
     <maven.compiler.target>21</maven.compiler.target>
